requirements.txt
asgiref==3.8.1
celery==5.3.6
Django==5.0.7
sqlparse==0.5.1
typing_extensions==4.12.2
redis==5.0.4
djangorestframework==3.15.1
requests==2.31.0
psycopg==3.1.18
Pillow==9.4.0
importlib-metadata==4.13.0